refactoring delete: better control
1. delItems are only handled in header part, not passed down to listitem
2. passed checkedItemHandler to listitem to reflect flag change on delete item list instead
